Refactoring test framework stuff
Before running the pipeline tests was dependent on a JVM variable, which seems kind of strange to me. So I have changed this to be a variable passed from the testng xml. However the value for runpipeline still has to be changed manually if one acutally wants to do more than dry run the tests. Hopefully this can be fixed in the future, but right now this will just have to do.