Fixing DNABestPracticeVariantCalling for non-human
This means some perhaps subtle changes. I moved to notHuman option into
the GATKConfig class as this is now used both in variant calling and
data processing. This idea is that now the FileAndProgramResourceConfig
trait will not load the default (human) resource files, and thus defer
them to what's been set on the command line. This should mean that the
human files are not loaded when it goes into the dataprocessing and thus
are not used.